refactor: move import to utils module
